+++ packages/react-dev-utils/package.json
@@ -11,13 +11,14 @@
     "node": ">=4"
   },
   "files": [
-    "clearConsole.js",
     "checkRequiredFiles.js",
+    "clearConsole.js",
+    "FileSizeReporter.js",
     "formatWebpackMessages.js",
     "getProcessForPort.js",
     "InterpolateHtmlPlugin.js",
-    "openChrome.applescript",
     "openBrowser.js",
+    "openChrome.applescript",
     "prompt.js",
     "WatchMissingNodeModulesPlugin.js",
     "webpackHotDevClient.js"
@@ -26,8 +27,11 @@
     "ansi-html": "0.0.5",
     "chalk": "1.1.3",
     "escape-string-regexp": "1.0.5",
+    "filesize": "3.3.0",
+    "gzip-size": "3.0.0",
     "html-entities": "1.2.0",
     "opn": "4.0.2",
+    "recursive-readdir": "2.1.1",
     "sockjs-client": "1.1.2",
     "strip-ansi": "3.0.1"
   }
