+++ packages/react-scripts/package.json
@@ -46,9 +46,7 @@
     "eslint-plugin-react": "6.4.1",
     "extract-text-webpack-plugin": "2.0.0",
     "file-loader": "0.10.0",
-    "filesize": "3.3.0",
     "fs-extra": "0.30.0",
-    "gzip-size": "3.0.0",
     "html-webpack-plugin": "2.28.0",
     "http-proxy-middleware": "0.17.3",
     "jest": "18.1.0",
@@ -56,8 +54,6 @@
     "postcss-loader": "1.3.1",
     "promise": "7.1.1",
     "react-dev-utils": "^0.5.1",
-    "recursive-readdir": "2.1.1",
-    "strip-ansi": "3.0.1",
     "style-loader": "0.13.1",
     "url-loader": "0.5.7",
     "webpack": "2.2.1",
